Moved ContainerNode to its own file
