+++ lluvia/cpp/core/src/CommandBuffer.cpp
@@ -2,7 +2,7 @@
 
 #include "lluvia/core/buffer/Buffer.h"
 #include "lluvia/core/node/ComputeNode.h"
-#include "lluvia/core/ContainerNode.h"
+#include "lluvia/core/node/ContainerNode.h"
 #include "lluvia/core/Duration.h"
 #include "lluvia/core/image/Image.h"
 #include "lluvia/core/image/ImageView.h"
