+++ lluvia/cpp/core/src/node/ContainerNode.cpp
@@ -1,11 +1,11 @@
 /**
 @file       ContainerNode.cpp
 @brief      ContainerNode class.
-@copyright  2018, Juan David Adarve Bermudez. See AUTHORS for more details.
+@copyright  2022, Juan David Adarve Bermudez. See AUTHORS for more details.
             Distributed under the Apache-2 license, see LICENSE for more details.
 */
 
-#include "lluvia/core/ContainerNode.h"
+#include "lluvia/core/node/ContainerNode.h"
 
 #include "lluvia/core/CommandBuffer.h"
 #include "lluvia/core/Interpreter.h"
