+++ lluvia/python/src/BUILD.bazel
@@ -13,8 +13,6 @@ pyx_library (
         "lluvia/core/core_object.pxd",
         "lluvia/core/duration.pxd",
         "lluvia/core/duration.pyx",
-        "lluvia/core/node.pxd",
-        "lluvia/core/node.pyx",
         "lluvia/core/parameter.pxd",
         "lluvia/core/parameter.pyx",
         "lluvia/core/program.pxd",
