+++ lluvia/python/src/lluvia/core/__init__.py
@@ -14,7 +14,6 @@ from .enums import *
 from .image import *
 from .memory import *
 from .node_new import *
-from .node import *
 from .parameter import *
 from .program import *
 from .session import *
