+++ lluvia/python/src/lluvia/core/command_buffer.pyx
@@ -18,8 +18,8 @@ from lluvia.core.duration cimport Duration
 from lluvia.core.image.image_layout cimport _ImageLayout, ImageLayout
 from lluvia.core.image.image cimport Image
 
-from lluvia.core.node cimport ContainerNode
 from lluvia.core.node_new.compute_node cimport ComputeNode
+from lluvia.core.node_new.container_node cimport ContainerNode
 
 
 __all__ = ['CommandBuffer']
