+++ lluvia/python/src/lluvia/core/session.pxd
@@ -23,11 +23,11 @@ from lluvia.core.duration cimport _Duration
 
 from lluvia.core.node_new.compute_node cimport _ComputeNode
 from lluvia.core.node_new.compute_node_descriptor cimport _ComputeNodeDescriptor
-from lluvia.core.node_new.node_builder_descriptor cimport _NodeBuilderDescriptor
 from lluvia.core.node_new.container_node_descriptor cimport _ContainerNodeDescriptor
+from lluvia.core.node_new.node_builder_descriptor cimport _NodeBuilderDescriptor
+from lluvia.core.node_new.container_node cimport _ContainerNode
 
 from lluvia.core.device.device_descriptor cimport _DeviceDescriptor
-from lluvia.core.node cimport _ContainerNode
 from lluvia.core.program cimport _Program
 from lluvia.core.types cimport _vec3ui
